Moved segment creation to single place
